pantalla tutorial finalizada
