+++ CashCarp/app/src/main/AndroidManifest.xml
@@ -13,17 +13,25 @@
         android:theme="@style/Theme.CashCarp"
         tools:targetApi="31">
         <activity
-            android:name=".mainscreen"
+            android:name=".TutorialScreen"
             android:exported="false" />
+
         <activity
             android:name=".MainActivity"
-            android:exported="true">
+            android:exported="false" />
+
+
+        <activity
+            android:name=".SplashActivity"
+            android:exported="true"
+            android:theme="@style/Theme.Design.NoActionBar">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+
     </application>
 
 </manifest>
