+++ CashCarp/app/src/main/java/com/proyecto/cashcarp/CreateAccountActivity.java
@@ -1,0 +1,14 @@
+package com.proyecto.cashcarp;
+
+import android.os.Bundle;
+import android.os.PersistableBundle;
+
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
+
+public class CreateAccountActivity extends AppCompatActivity {
+    @Override
+    public void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+    }
+}
