+++ CashCarp/app/src/main/java/com/proyecto/cashcarp/SplashActivity.java
@@ -1,0 +1,24 @@
+package com.proyecto.cashcarp;
+
+import android.content.Intent;
+import android.os.Bundle;
+import android.os.Handler;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+public class SplashActivity extends AppCompatActivity {
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.splash_screen);
+
+        new Handler().postDelayed(new Runnable() {
+            @Override
+            public void run() {
+                Intent i = new Intent(SplashActivity.this, TutorialScreen.class);
+                startActivity(i);
+                finish();
+            }
+        }, 3000);
+    }
+}
