+++ CashCarp/app/src/main/java/com/proyecto/cashcarp/ViewPagerAdapter.java
@@ -13,27 +13,28 @@ import androidx.viewpager.widget.PagerAdapter;
 
 public class ViewPagerAdapter extends PagerAdapter {
     Context context;
-    int images[] = {
+    int[] images = {
             R.drawable.foto_tutorial_1,
             R.drawable.foto_tutorial_2,
-            R.drawable.foto_tutorial_3,
+            R.drawable.foto_tutorial_3
     };
 
-    int titulo[] = {
+    int[] titulo = {
             R.string.titulo_tutorial_1,
             R.string.titulo_tutorial_2,
             R.string.titulo_tutorial_3
     };
 
-    int descripcion[] = {
+    int[] descripcion = {
             R.string.descripcion_tutorial_1,
             R.string.descripcion_tutorial_2,
             R.string.descripcion_tutorial_3
     };
 
-    public  ViewPagerAdapter(Context context){
+    public ViewPagerAdapter(Context context) {
         this.context = context;
     }
+
     @Override
     public int getCount() {
         return images.length;
@@ -47,13 +48,12 @@ public class ViewPagerAdapter extends PagerAdapter {
     @NonNull
     @Override
     public Object instantiateItem(@NonNull ViewGroup container, int position) {
+        LayoutInflater layoutInflater = (LayoutInflater) context.getSystemService(Context.LAYOUT_INFLATER_SERVICE);
+        View view = layoutInflater.inflate(R.layout.slider_layout, container, false);
 
-        LayoutInflater layoutInflater = (LayoutInflater) context.getSystemService(context.LAYOUT_INFLATER_SERVICE);
-        View view = layoutInflater.inflate(R.layout.slider_layout,container,false);
-
-        ImageView slideTutorialImage = (ImageView) view.findViewById(R.id.tutorialImage);
-        TextView slideTutorialTitle = (TextView) view.findViewById(R.id.tutorialTitle);
-        TextView slideTutorialDesc = (TextView)  view.findViewById(R.id.tutorialDescription);
+        ImageView slideTutorialImage = view.findViewById(R.id.tutorialImage);
+        TextView slideTutorialTitle = view.findViewById(R.id.tutorialTitle);
+        TextView slideTutorialDesc = view.findViewById(R.id.tutorialDescription);
 
         slideTutorialImage.setImageResource(images[position]);
         slideTutorialTitle.setText(titulo[position]);
@@ -62,8 +62,8 @@ public class ViewPagerAdapter extends PagerAdapter {
         container.addView(view);
 
         return view;
-
     }
+
     @Override
     public void destroyItem(@NonNull ViewGroup container, int position, @NonNull Object object) {
         container.removeView((View) object);
