--- CashCarp/app/src/main/java/com/proyecto/cashcarp/mainscreen.java
@@ -1,14 +1,0 @@
-package com.proyecto.cashcarp;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.os.Bundle;
-
-public class mainscreen extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_mainscreen);
-    }
-}
