--- CashCarp/app/src/main/res/layout/fragment_splash_screen.xml
@@ -1,26 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="@color/material_dynamic_neutral60"
-    tools:context=".SplashScreen">
-
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@string/app_name"
-        android:textSize="65sp"
-        android:layout_marginTop="52dp"
-        android:layout_gravity="center_horizontal"
-        android:textColor="@color/black" />
-
-    <ImageView
-        android:id="@+id/miImagen"
-        android:layout_width="321dp"
-        android:layout_height="545dp"
-        android:src="@drawable/cashcarp_transparente"
-        android:layout_gravity="center"/>
-
-
-</FrameLayout>
